+++ neo/tools/radiant/cmdlib.cpp
@@ -207,7 +207,7 @@ void StripFilename (char *path)
 {
 	int             length;
 
-	length = strlen(path)-1;
+	length = (int)strlen(path)-1;
 	while (length > 0 && path[length] != PATHSEPERATOR) {
 		length--;
 	}
@@ -223,7 +223,7 @@ void StripExtension (char *path)
 {
 	int             length;
 
-	length = strlen(path)-1;
+	length = (int)strlen(path)-1;
 	while (length > 0 && path[length] != '.')
 	{
 		length--;
